wasmer: start of implementation
