+++ packages/wasm/Cargo.toml
@@ -6,7 +6,6 @@ description = "Todo"
 edition = "2021"
 
 [dependencies]
-wasmtime = "1.0.2"
 rmp-serde = "1.1.1"
 rmpv = "1.0.0"
 serde = "1.0.145"
@@ -15,6 +14,7 @@ async-trait = "0.1.57"
 base64 = "0.13.0"
 futures = "0.3.24"
 serde_json = "1.0.87"
+wasmer = "3.1.1"
 polywrap_core = {path = "../core"}
 wrap_manifest_schemas = {path = "../manifest"}
 polywrap_msgpack = {path = "../msgpack"}
