+++ packages/wasm/src/wasm_runtime/mod.rs
@@ -1,2 +1,4 @@
+pub mod old_imports;
+pub mod old_instance;
 pub mod imports;
 pub mod instance;
